Adds various framework for the pulling and pushing of data to and from the activity.
Adds various framework for the pulling and pushing of data to and from the activity.For get requests in the gameactivity, pullgamedatathread, and signuploginact putting quiryperameters notifying what kind of get request this is the placeholder. For push requests, a JSON object should be put in the body of the request and from the JSON object the server should react based on what the object has within it. For the playerupdates, data will be obtianed in a thread and put into the gameusers ArrayMap. This arraymap maps the player's name to the player object that has the player's username, Latitude, Longitude, and HumOrZomb string denoting wheather they are "Human" or "Zombie". Some miliseconds (right now 500) after the thread is done, an empty message will be sent to the
peopleupdate handler in GameActivity signifying that the pullgamedatathread is done and the ArrayMap is free to be used to update the gamestate. At the end of the handler the pullgamedatathread is created and started again.